move models in the models module + use a proper appconfig rather than hacky way in models.py previously.
