+++ django_cassandra_engine/__init__.py
@@ -6,3 +6,5 @@ __author__ = "Rafał Furmański"
 __contact__ = "r.furmanski@gmail.com"
 __homepage__ = "http://github.com/r4fek/django-cassandra-engine"
 __docformat__ = "markdown"
+
+default_app_config = 'django_cassandra_engine.apps.AppConfig'
