+++ requirements.txt
@@ -1,3 +1,3 @@
 cassandra-driver>=3.3.0,<3.4.2
-Django<1.10
+Django>=1.9,<1.10
 six>=1.6
